+++ src/main/java/com/bookmile/backend/domain/review/service/ReviewService.java
@@ -1,70 +1,16 @@
 package com.bookmile.backend.domain.review.service;
 
-import static com.bookmile.backend.global.common.StatusCode.BOOK_NOT_FOUND;
-import static com.bookmile.backend.global.common.StatusCode.REVIEW_NOT_FOUND;
-import static com.bookmile.backend.global.common.StatusCode.USER_NOT_FOUND;
-
-import com.bookmile.backend.domain.book.entity.Book;
 import com.bookmile.backend.domain.review.dto.req.ReviewReqDto;
 import com.bookmile.backend.domain.review.dto.res.ReviewListResDto;
-import com.bookmile.backend.domain.review.entity.Review;
-import com.bookmile.backend.domain.review.repository.ReviewRepository;
-import com.bookmile.backend.domain.user.entity.User;
-import com.bookmile.backend.domain.user.repository.UserRepository;
-import com.bookmile.backend.global.exception.CustomException;
 import java.util.List;
-import java.util.stream.Collectors;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
-@Service
-@RequiredArgsConstructor
-public class ReviewService {
-    private final ReviewRepository reviewRepository;
-    private final BookRepository bookRepository;
-    private final UserRepository userRepository;
+public interface ReviewService {
+    List<ReviewListResDto> viewReviewList(Long bookId);
 
-    public List<ReviewListResDto> viewReviewList(Long bookId) {
-        Book book = bookRepository.findById(bookId)
-                .orElseThrow(() -> new CustomException(BOOK_NOT_FOUND));
+    Long createReview(Long bookId, Long userId, ReviewReqDto reviewReqDto);
 
-        return reviewRepository.findAllByBookId(bookId).stream()
-                .map(ReviewListResDto::createReview)
-                .collect(Collectors.toList());
-    }
+    Long updateReview(Long reviewId, ReviewReqDto reviewReqDto);
 
-    public Long createReview(Long bookId, Long userId, ReviewReqDto reviewReqDto) {
-        Book book = bookRepository.findById(bookId)
-                .orElseThrow(() -> new CustomException(BOOK_NOT_FOUND));
-
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new CustomException(USER_NOT_FOUND));
-
-        Review review = Review.from(user, book, reviewReqDto);
-
-        reviewRepository.save(review);
-
-        return review.getId();
-    }
-
-    public Long updateReview(Long reviewId, ReviewReqDto reviewReqDto) {
-        Review review = reviewRepository.findById(reviewId)
-                .orElseThrow(() -> new CustomException(REVIEW_NOT_FOUND));
-
-        review.update(reviewReqDto);
-        reviewRepository.save(review);
-
-        return review.getId();
-    }
-
-    public Long deleteReview(Long reviewId) {
-        Review review = reviewRepository.findById(reviewId)
-                .orElseThrow(() -> new CustomException(REVIEW_NOT_FOUND));
-
-        review.delete(review);
-
-        reviewRepository.save(review);
-
-        return review.getId();
-    }
+    Long deleteReview(Long reviewId);
 }
+
